feature/WA-14 Refactored CityPage
