+++ src/components/pages/CityPage/CityPage.module.scss
@@ -1,6 +1,5 @@
 .main {
   background-color: #d7a7cf;
-  width: 864px;
   height: 900px;
   display: flex;
   align-items: center;
@@ -49,10 +48,6 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-}
-
-.separator {
-  margin: 0;
 }
 
 .forecastBlock {
