--- src/components/pages/CityPage/CurrentWeatherElement/index.js
@@ -1,1 +1,0 @@
-export { default as CurrentWeatherElement } from './CurrentWeatherElement';
