--- src/components/pages/CityPage/ForecastElement/index.js
@@ -1,1 +1,0 @@
-export { default as ForecastElement } from './ForecastElement';
